+++ src/main/java/edu/ie3/datamodel/utils/ContainerUtils.java
@@ -5,6 +5,12 @@
 */
 package edu.ie3.datamodel.utils;
 
+import static edu.ie3.util.quantities.PowerSystemUnits.KILOMETRE;
+import static edu.ie3.util.quantities.PowerSystemUnits.OHM_PER_KILOMETRE;
+import static java.lang.Math.pow;
+import static java.lang.Math.sqrt;
+import static tech.units.indriya.unit.Units.OHM;
+
 import edu.ie3.datamodel.exceptions.InvalidGridException;
 import edu.ie3.datamodel.exceptions.TopologyException;
 import edu.ie3.datamodel.graph.*;
@@ -16,13 +22,16 @@ import edu.ie3.datamodel.models.input.graphics.LineGraphicInput;
 import edu.ie3.datamodel.models.input.graphics.NodeGraphicInput;
 import edu.ie3.datamodel.models.input.system.*;
 import edu.ie3.datamodel.models.voltagelevels.VoltageLevel;
+import edu.ie3.util.quantities.interfaces.SpecificResistance;
 import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import javax.measure.quantity.Length;
 import org.jgrapht.graph.DirectedMultigraph;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import tech.units.indriya.ComparableQuantity;
 
 /** Offers functionality useful for grouping different models together */
 public class ContainerUtils {
@@ -129,6 +138,149 @@ public class ContainerUtils {
     graph.addEdge(nodeA, nodeB);
     graph.setEdgeWeight(
         graph.getEdge(nodeA, nodeB), GridAndGeoUtils.distanceBetweenNodes(nodeA, nodeB));
+  }
+
+  /**
+   * Returns the topology of the provided grid container as a {@link ImpedanceWeightedGraph} if the
+   * provided grid container's {@link RawGridElements} allows the creation of a valid topology graph
+   * or an empty optional otherwise.
+   *
+   * @param grid the grid container that should be converted into topology graph
+   * @return either an optional holding the impedance topology graph instance or an empty optional
+   */
+  public static Optional<ImpedanceWeightedGraph> getImpedanceTopologyGraph(GridContainer grid) {
+    return getImpedanceTopologyGraph(grid.getRawGrid());
+  }
+
+  /**
+   * Returns the topology of the provided {@link RawGridElements} as a {@link
+   * ImpedanceWeightedGraph}, if they allow the creation of a valid topology graph or an empty
+   * optional otherwise.
+   *
+   * @param rawGridElements raw grids elements as base of the distance weighted topology graph
+   * @return either an optional holding the impedance topology graph instance or an empty optional
+   */
+  public static Optional<ImpedanceWeightedGraph> getImpedanceTopologyGraph(
+      RawGridElements rawGridElements) {
+
+    ImpedanceWeightedGraph graph = new ImpedanceWeightedGraph();
+
+    try {
+      rawGridElements.getNodes().forEach(graph::addVertex);
+    } catch (NullPointerException ex) {
+      log.error("At least one node entity of provided RawGridElements is null. ", ex);
+      return Optional.empty();
+    }
+
+    try {
+      rawGridElements.getLines().forEach(line -> addImpedanceGraphEdge(graph, line));
+    } catch (NullPointerException | IllegalArgumentException | UnsupportedOperationException ex) {
+      log.error("Error adding line edges to graph: ", ex);
+      return Optional.empty();
+    }
+
+    try {
+      rawGridElements
+          .getSwitches()
+          .forEach(
+              switchInput -> {
+                if (switchInput.isClosed()) addImpedanceGraphEdge(graph, switchInput);
+              });
+    } catch (NullPointerException | IllegalArgumentException | UnsupportedOperationException ex) {
+      log.error("Error adding switch edges to graph: ", ex);
+      return Optional.empty();
+    }
+
+    try {
+      rawGridElements.getTransformer2Ws().forEach(trafo2w -> addImpedanceGraphEdge(graph, trafo2w));
+    } catch (NullPointerException | IllegalArgumentException | UnsupportedOperationException ex) {
+      log.error("Error adding 2 winding transformer edges to graph: ", ex);
+      return Optional.empty();
+    }
+
+    try {
+      rawGridElements.getTransformer3Ws().forEach(trafo3w -> addImpedanceGraphEdge(graph, trafo3w));
+    } catch (NullPointerException | IllegalArgumentException | UnsupportedOperationException ex) {
+      log.error("Error adding 3 winding transformer edges to graph: ", ex);
+      return Optional.empty();
+    }
+
+    // if we reached this point, we can safely return a valid graph
+    return Optional.of(graph);
+  }
+
+  /**
+   * Adds an {@link ImpedanceWeightedEdge} to the provided graph between the provided nodes a and b.
+   * By implementation of jGraphT this side effect cannot be removed. :(
+   *
+   * @param graph the graph to be altered
+   * @param connectorInput the connector input element
+   */
+  private static void addImpedanceGraphEdge(
+      ImpedanceWeightedGraph graph, ConnectorInput connectorInput) {
+    NodeInput nodeA = connectorInput.getNodeA();
+    NodeInput nodeB = connectorInput.getNodeB();
+    /* Add an edge if it is not a switch or the switch is closed */
+    if (!(connectorInput instanceof SwitchInput) || ((SwitchInput) connectorInput).isClosed())
+      graph.addEdge(nodeA, nodeB);
+
+    if (connectorInput instanceof LineInput) {
+      LineInput line = (LineInput) connectorInput;
+      graph.setEdgeWeight(
+          graph.getEdge(nodeA, nodeB),
+          calcImpedance(line.getType().getR(), line.getType().getX(), line.getLength()));
+    }
+    if (connectorInput instanceof SwitchInput) {
+      SwitchInput sw = (SwitchInput) connectorInput;
+      // assumption: closed switch has a resistance of 1 OHM
+      if (sw.isClosed()) graph.setEdgeWeight(graph.getEdge(nodeA, nodeB), 1);
+    }
+    if (connectorInput instanceof Transformer2WInput) {
+      Transformer2WInput trafo2w = (Transformer2WInput) connectorInput;
+      graph.setEdgeWeight(
+          graph.getEdge(nodeA, nodeB),
+          sqrt(
+              pow(trafo2w.getType().getrSc().to(OHM).getValue().doubleValue(), 2)
+                  + pow(trafo2w.getType().getxSc().to(OHM).getValue().doubleValue(), 2)));
+    }
+    if (connectorInput instanceof Transformer3WInput) {
+      Transformer3WInput trafo3w = (Transformer3WInput) connectorInput;
+      graph.addEdge(nodeA, trafo3w.getNodeC());
+
+      graph.setEdgeWeight(
+          graph.getEdge(nodeA, nodeB),
+          sqrt(
+              pow(
+                      trafo3w.getType().getrScA().to(OHM).getValue().doubleValue()
+                          + trafo3w.getType().getrScB().to(OHM).getValue().doubleValue(),
+                      2)
+                  + pow(
+                      trafo3w.getType().getxScA().to(OHM).getValue().doubleValue()
+                          + trafo3w.getType().getxScB().to(OHM).getValue().doubleValue(),
+                      2)));
+
+      graph.setEdgeWeight(
+          graph.getEdge(nodeA, trafo3w.getNodeC()),
+          sqrt(
+              pow(
+                      trafo3w.getType().getrScA().to(OHM).getValue().doubleValue()
+                          + trafo3w.getType().getrScC().to(OHM).getValue().doubleValue(),
+                      2)
+                  + pow(
+                      trafo3w.getType().getxScA().to(OHM).getValue().doubleValue()
+                          + trafo3w.getType().getxScC().to(OHM).getValue().doubleValue(),
+                      2)));
+    }
+  }
+
+  private static double calcImpedance(
+      ComparableQuantity<SpecificResistance> r,
+      ComparableQuantity<SpecificResistance> x,
+      ComparableQuantity<Length> length) {
+    return sqrt(
+            pow(r.to(OHM_PER_KILOMETRE).getValue().doubleValue(), 2)
+                + pow(x.to(OHM_PER_KILOMETRE).getValue().doubleValue(), 2))
+        * length.to(KILOMETRE).getValue().doubleValue();
   }
 
   /**